+++ app/controllers/StripeController.scala
@@ -1,17 +1,153 @@
 package controllers
 
+import java.lang.Math._
+
 import actions.CommonActions._
-import cats.data.Xor
+import cats.data.{Xor, XorT}
+import com.gu.i18n.CountryGroup._
+import com.gu.i18n.{AUD, Currency, EUR, USD}
+import com.gu.stripe.Stripe
+import com.gu.stripe.Stripe.Charge
+import com.gu.stripe.Stripe.Serializer._
 import com.typesafe.config.Config
-import models.StripeHook
+import models.{ContributionId, IdentityId, SavedContributionData, StripeHook}
+import org.joda.time.DateTime
 import play.api.Logger
-import play.api.libs.json.{JsError, JsSuccess, JsValue}
+import play.api.data.{Form, FormError}
+import play.api.data.Forms._
+import play.api.data.format.Formatter
+import play.api.libs.json.{JsError, JsSuccess, JsValue, Json}
 import play.api.mvc.{BodyParsers, Controller, Result}
 import services.PaymentServices
+import utils.MaxAmount
+import views.support.Test
 
 import scala.concurrent.{ExecutionContext, Future}
 
 class StripeController(paymentServices: PaymentServices, stripeConfig: Config)(implicit ec: ExecutionContext) extends Controller {
+
+  implicit val currencyFormatter = new Formatter[Currency] {
+    type Result = Either[Seq[FormError], Currency]
+
+    override def bind(key: String, data: Map[String, String]): Result =
+      data.get(key).map(_.toUpperCase).flatMap(Currency.fromString).fold[Result](Left(Seq.empty))(currency => Right(currency))
+
+    override def unbind(key: String, value: Currency): Map[String, String] =
+      Map(key -> value.identifier)
+  }
+
+  case class JsonAbTest(testName: String, testSlug: String, variantName: String, variantSlug: String)
+
+  object JsonAbTest {
+    implicit val abTestFormat = Json.format[JsonAbTest]
+  }
+
+  case class SupportForm(
+    name: String,
+    currency: Currency,
+    amount: BigDecimal,
+    email: String,
+    token: String,
+    marketing: Boolean,
+    postcode: Option[String],
+    abTests: Set[JsonAbTest],
+    ophanPageviewId: String,
+    ophanBrowserId: Option[String],
+    cmp: Option[String],
+    intcmp: Option[String]
+
+  )
+
+  val supportForm: Form[SupportForm] = Form(
+    mapping(
+      "name" -> text,
+      "currency" -> of[Currency],
+      "amount" -> bigDecimal(10, 2),
+      "email" -> email,
+      "token" -> nonEmptyText,
+      "marketing" -> boolean,
+      "postcode" -> optional(nonEmptyText),
+      "abTests" -> set(mapping(
+        "testName" -> text,
+        "testSlug" -> text,
+        "variantName" -> text,
+        "variantSlug" -> text
+      )(JsonAbTest.apply)(JsonAbTest.unapply)),
+      "ophanPageviewId" -> text,
+      "ophanBrowserId" -> optional(text),
+      "cmp" -> optional(text),
+      "intcmp" -> optional(text)
+    )(SupportForm.apply)(SupportForm.unapply)
+  )
+
+  def pay = NoCacheAction.async(parse.form(supportForm)) { implicit request =>
+
+    val form = request.body
+
+    val stripe = paymentServices.stripeServiceFor(request)
+    val idUser = IdentityId.fromRequest(request)
+
+    val countryGroup = form.currency match {
+      case USD => US
+      case AUD => Australia
+      case EUR => Europe
+      case _ => UK
+    }
+
+    val variant = Test.getContributePageVariant(countryGroup, Test.testIdFor(request), request)
+
+    val contributionId = ContributionId.random
+
+    val metadata = Map(
+      "marketing-opt-in" -> form.marketing.toString,
+      "email" -> form.email,
+      "name" -> form.name,
+      "abTests" -> Json.toJson(Seq(variant)).toString,
+      "ophanPageviewId" -> form.ophanPageviewId,
+      "ophanBrowserId" -> form.ophanBrowserId.getOrElse(""),
+      "cmp" -> form.cmp.mkString,
+      "intcmp" -> form.intcmp.mkString,
+      "contributionId" -> contributionId.toString
+    ) ++ List(
+      form.postcode.map("postcode" -> _),
+      idUser.map("idUser" -> _.id)
+    ).flatten.toMap
+    // Note that '.. * 100' will not work for Yen and other currencies! https://stripe.com/docs/api#charge_object-amount
+    val amountInSmallestCurrencyUnit = (form.amount * 100).toInt
+    val maxAmountInSmallestCurrencyUnit = MaxAmount.forCurrency(form.currency) * 100
+    val amount = min(maxAmountInSmallestCurrencyUnit, amountInSmallestCurrencyUnit)
+
+    val redirect = routes.Giraffe.thanks(countryGroup).url
+
+    def createCharge: Future[Stripe.Charge] = {
+      stripe.Charge.create(amount, form.currency, form.email, "Your contribution", form.token, metadata)
+    }
+
+    def storeMetaData(charge: Charge): XorT[Future, String, SavedContributionData] = {
+      stripe.storeMetaData(
+        contributionId = contributionId,
+        created = new DateTime(charge.created * 1000L),
+        email = charge.receipt_email,
+        name = form.name,
+        postCode = form.postcode,
+        marketing = form.marketing,
+        variants = Seq(variant),
+        cmp = form.cmp,
+        intCmp = form.intcmp,
+        ophanPageviewId = form.ophanPageviewId,
+        ophanBrowserId = form.ophanBrowserId,
+        idUser = idUser
+      )
+    }
+
+    createCharge.map { charge =>
+      storeMetaData(charge) // fire and forget. If it fails we don't want to stop the user
+      Ok(Json.obj("redirect" -> redirect))
+        .withSession("charge_id" -> charge.id)
+    }.recover {
+      case e: Stripe.Error => BadRequest(Json.toJson(e))
+    }
+  }
 
   val webhookKey = stripeConfig.getString("webhook.key")
